Remove AccountListModel
