+++ operations/AccountComboBox.qml
@@ -2,8 +2,8 @@ import QtQuick.Controls
 
 ComboBox {
     required property var budgetData
-    model: budgetData.accountModel
+    model: budgetData
     enabled: budgetData.accountCount > 0
     textRole: "name"
-    property var currentAccount: budgetData.accountModel.accountAt(currentIndex)
+    property var currentAccount: budgetData.accountAt(currentIndex)
 }
